requirements.txt
pydantic
requests
pyyaml
google-generativeai
